SeeFoodServer/requirements.txt
fastapi
uvicorn
torch==2.0.1+cpu
torchvision==0.15.2+cpu
pillow
pydantic
